+++ dev4press/core/options/Render.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Name:    Dev4Press\v46\Core\Options\Render
- * Version: v4.6
+ * Name:    Dev4Press\v47\Core\Options\Render
+ * Version: v4.7
  * Author:  Milan Petrovic
  * Email:   support@dev4press.com
  * Website: https://www.dev4press.com/
@@ -25,13 +25,12 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-namespace Dev4Press\v46\Core\Options;
+namespace Dev4Press\v47\Core\Options;
 
-use Dev4Press\v46\Core\Quick\Arr;
-use Dev4Press\v46\Core\Quick\KSES;
-use Dev4Press\v46\Core\Quick\Sanitize;
-use Dev4Press\v46\Core\UI\Elements;
-use Dev4Press\v46\WordPress\Walker\CheckboxRadio;
+use Dev4Press\v47\Core\Quick\Arr;
+use Dev4Press\v47\Core\Quick\KSES;
+use Dev4Press\v47\Core\Quick\Sanitize;
+use Dev4Press\v47\Core\UI\Elements;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -367,15 +366,28 @@ class Render {
 		echo '</div>';
 	}
 
-	protected function _pair_element( $name, $id, $i, $value, $element, $hide = false ) {
-		echo '<div class="pair-element-' . esc_attr( $i ) . '" style="display: ' . ( $hide ? 'none' : 'block' ) . '">';
-		echo '<label for="' . esc_attr( $id ) . '_key">' . KSES::strong( $element->args['label_key'] ) . ':</label>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-		echo '<input type="text" name="' . esc_attr( $name ) . '[key]" id="' . esc_attr( $id ) . '_key" value="' . esc_attr( $value['key'] ) . '" class="widefat" />';
+	protected function _pair_element( $name, $id, $i, $value, $element, $hide = false, $layout = 'normal' ) {
+		$type_key    = $element->args['type_key'] ?? 'text';
+		$type_value  = $element->args['type_value'] ?? 'text';
+		$label_key   = $element->args['label_key'] ?? __( 'Key', 'd4plib' );
+		$label_value = $element->args['label_value'] ?? __( 'Value', 'd4plib' );
+		$remove      = $element->args['label_button_remove'] ?? '<i aria-label="' . __( 'Remove', 'd4plib' ) . '" class="d4p-icon d4p-ui-times"></i><span class="d4p-sr-only">' . __( 'Remove', 'd4plib' ) . '</span>';
 
-		echo '<label for="' . esc_attr( $id ) . '_value">' . KSES::strong( $element->args['label_value'] ) . ':</label>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-		echo '<input type="text" name="' . esc_attr( $name ) . '[value]" id="' . esc_attr( $id ) . '_value" value="' . esc_attr( $value['value'] ) . '" class="widefat" />';
+		echo '<div class="d4p-element-pair-single pair-element-' . esc_attr( $i ) . '" style="' . ( $hide ? 'display:none' : '' ) . '">';
 
-		echo '<a role="button" class="button-secondary" href="#">' . KSES::strong( $element->args['label_button_remove'] ) . '</a>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		if ( $layout == 'normal' ) {
+			echo '<label for="' . esc_attr( $id ) . '_key">' . KSES::strong( $label_key ) . ':</label>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		}
+
+		echo '<div><input placeholder="' . esc_attr( $label_key ) . '" type="' . esc_attr( $type_key ) . '" name="' . esc_attr( $name ) . '[key]" id="' . esc_attr( $id ) . '_key" value="' . esc_attr( $value['key'] ) . '" class="widefat" /></div>';
+
+		if ( $layout == 'normal' ) {
+			echo '<label for="' . esc_attr( $id ) . '_value">' . KSES::strong( $label_value ) . ':</label>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		}
+
+		echo '<div><input placeholder="' . esc_attr( $label_value ) . '" type="' . esc_attr( $type_value ) . '" name="' . esc_attr( $name ) . '[value]" id="' . esc_attr( $id ) . '_value" value="' . esc_attr( $value['value'] ) . '" class="widefat" /></div>';
+
+		echo '<span><a role="button" class="button-secondary" href="#">' . KSES::strong( $remove ) . '</a></span>'; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		echo '</div>';
 	}
 
@@ -393,8 +405,8 @@ class Render {
 
 	protected function _datetime_element( Element $element, $value, $name_base, $id_base, $type = 'text', $class = '' ) {
 		$readonly  = isset( $element->args['readonly'] ) && $element->args['readonly'] ? ' readonly' : '';
-		$min       = isset( $element->args['min'] ) ? ' min="' . esc_attr($element->args['min']) . '"' : '';
-		$max       = isset( $element->args['max'] ) ? ' max="' . esc_attr($element->args['max']) . '"' : '';
+		$min       = isset( $element->args['min'] ) ? ' min="' . esc_attr( $element->args['min'] ) . '"' : '';
+		$max       = isset( $element->args['max'] ) ? ' max="' . esc_attr( $element->args['max'] ) . '"' : '';
 		$flatpickr = isset( $element->args['flatpickr'] ) && $element->args['flatpickr'];
 		$type      = $flatpickr ? 'text' : $type;
 		$class     = 'widefat' . ( $flatpickr ? ' ' . $class : '' );
@@ -429,59 +441,36 @@ class Render {
 		$this->_datetime_element( $element, $value, $name_base, $id_base, 'datetime-local', 'd4p-input-field-datetime' );
 	}
 
-	protected function draw_license( Element $element, $value, $name_base, $id_base, $type = 'text', $class = 'widefat' ) {
-		$readonly    = isset( $element->args['readonly'] ) && $element->args['readonly'] ? ' readonly' : '';
-		$placeholder = isset( $element->args['placeholder'] ) && ! empty( $element->args['placeholder'] ) ? $element->args['placeholder'] : '';
-		$type        = 'text';
-		$pattern     = '^\d{4}-\d{8}-[A-Z0-9]{6}-[A-Z0-9]{6}-\d{4}$';
-
-		echo sprintf(
-			'<input aria-labelledby="%s__label"%s placeholder="%s" pattern="%s" type="%s" name="%s" id="%s" value="%s" class="%s" />',
-			esc_attr( $id_base ),
-			esc_attr( $readonly ),
-			esc_attr( $placeholder ),
-			$pattern,
-			$type,
-			esc_attr( $name_base ),
-			esc_attr( $id_base ),
-			esc_attr( $value ),
-			esc_attr( $class )
-		);
+	protected function draw_license( Element $element, $value, $name_base, $id_base ) {
+		Elements::instance()->input( $value, array(
+			'echo'        => true,
+			'id'          => $id_base,
+			'name'        => $name_base,
+			'class'       => 'widefat',
+			'type'        => 'text',
+			'readonly'    => isset( $element->args['readonly'] ) && $element->args['readonly'],
+			'placeholder' => isset( $element->args['placeholder'] ) && ! empty( $element->args['placeholder'] ) ? $element->args['placeholder'] : '',
+			'pattern'     => '^\d{4}-\d{8}-[A-Z0-9]{6}-[A-Z0-9]{6}-\d{4}$',
+		), array(), array(
+			'labelledby' => $id_base . '__label',
+		) );
 	}
 
-	protected function draw_text( Element $element, $value, $name_base, $id_base, $type = 'text', $class = 'widefat' ) {
-		$readonly    = isset( $element->args['readonly'] ) && $element->args['readonly'] ? ' readonly' : '';
-		$placeholder = isset( $element->args['placeholder'] ) && ! empty( $element->args['placeholder'] ) ? $element->args['placeholder'] : '';
-		$type        = isset( $element->args['type'] ) && ! empty( $element->args['type'] ) ? $element->args['type'] : $type;
-
-		if ( isset( $element->args['pattern'] ) ) {
-			$pattern = $element->args['pattern'];
-
-			echo sprintf(
-				'<input aria-labelledby="%s__label"%s placeholder="%s" pattern="%s" type="%s" name="%s" id="%s" value="%s" class="%s" />',
-				esc_attr( $id_base ),
-				esc_attr( $readonly ),
-				esc_attr( $placeholder ),
-				esc_attr( $pattern ),
-				esc_attr( $type ),
-				esc_attr( $name_base ),
-				esc_attr( $id_base ),
-				esc_attr( $value ),
-				esc_attr( $class )
-			);
-		} else {
-			echo sprintf(
-				'<input aria-labelledby="%s__label"%s placeholder="%s" type="%s" name="%s" id="%s" value="%s" class="%s" />',
-				esc_attr( $id_base ),
-				esc_attr( $readonly ),
-				esc_attr( $placeholder ),
-				esc_attr( $type ),
-				esc_attr( $name_base ),
-				esc_attr( $id_base ),
-				esc_attr( $value ),
-				esc_attr( $class )
-			);
-		}
+	protected function draw_text( Element $element, $value, $name_base, $id_base, $type = 'text' ) {
+		Elements::instance()->input( $value, array(
+			'echo'        => true,
+			'id'          => $id_base,
+			'name'        => $name_base,
+			'class'       => 'widefat',
+			'minlength'   => $element->args['minlength'] ?? '',
+			'maxlength'   => $element->args['maxlength'] ?? '',
+			'type'        => isset( $element->args['type'] ) && ! empty( $element->args['type'] ) ? $element->args['type'] : $type,
+			'readonly'    => isset( $element->args['readonly'] ) && $element->args['readonly'],
+			'placeholder' => isset( $element->args['placeholder'] ) && ! empty( $element->args['placeholder'] ) ? $element->args['placeholder'] : '',
+			'pattern'     => isset( $element->args['pattern'] ) && ! empty( $element->args['pattern'] ) ? $element->args['pattern'] : '',
+		), array(), array(
+			'labelledby' => $id_base . '__label',
+		) );
 	}
 
 	protected function draw_html( Element $element, $value, $name_base, $id_base ) {
@@ -498,23 +487,20 @@ class Render {
 	}
 
 	protected function draw_number( Element $element, $value, $name_base, $id_base ) {
-		$readonly = isset( $element->args['readonly'] ) && $element->args['readonly'] ? ' readonly' : '';
-
-		$min  = isset( $element->args['min'] ) ? ' min="' . floatval( esc_attr( $element->args['min'] ) ) . '"' : '';
-		$max  = isset( $element->args['max'] ) ? ' max="' . floatval( esc_attr( $element->args['max'] ) ) . '"' : '';
-		$step = isset( $element->args['step'] ) ? ' step="' . floatval( esc_attr( $element->args['step'] ) ) . '"' : '';
-
-		echo sprintf(
-			'<input aria-labelledby="%s__label"%s type="number" name="%s" id="%s" value="%s" class="widefat"%s%s%s />',
-			esc_attr( $id_base ),
-			esc_attr( $readonly ),
-			esc_attr( $name_base ),
-			esc_attr( $id_base ),
-			esc_attr( $value ),
-			$min,
-			$max,
-			$step
-		);
+		Elements::instance()->input( $value, array(
+			'echo'        => true,
+			'id'          => $id_base,
+			'name'        => $name_base,
+			'class'       => 'widefat',
+			'type'        => 'number',
+			'min'         => $element->args['min'] ?? '',
+			'max'         => $element->args['max'] ?? '',
+			'step'        => $element->args['step'] ?? '',
+			'readonly'    => isset( $element->args['readonly'] ) && $element->args['readonly'],
+			'placeholder' => isset( $element->args['placeholder'] ) && ! empty( $element->args['placeholder'] ) ? $element->args['placeholder'] : '',
+		), array(), array(
+			'labelledby' => $id_base . '__label',
+		) );
 
 		if ( isset( $element->args['label_unit'] ) ) {
 			echo '<span class="d4p-field-unit">' . esc_html( $element->args['label_unit'] ) . '</span>';
@@ -716,7 +702,6 @@ class Render {
 		$child      = $element->args['child_of'] ?? 0;
 		$depth      = $element->args['depth'] ?? 0;
 
-		// phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		$list = wp_dropdown_pages(
 			array(
 				'echo'              => false,
@@ -1022,6 +1007,13 @@ class Render {
 	}
 
 	protected function draw_expandable_pairs( Element $element, $value, $name_base, $id_base = '' ) {
+		$layout = $element->args['layout'] ?? 'normal';
+		$class  = array(
+			'd4p-expandable-pairs-wrapper',
+			'd4p-expandable-pairs-layout-' . $layout,
+		);
+
+		echo '<div class="' . Sanitize::html_classes( $class ) . '">';
 		$this->_pair_element(
 			$name_base . '[0]',
 			$id_base . '_0',
@@ -1031,8 +1023,22 @@ class Render {
 				'value' => '',
 			),
 			$element,
-			true
+			true,
+			$layout
 		);
+
+		echo '<div class="d4p-expandable-pairs-inside">';
+
+		if ( $layout == 'inline' ) {
+			$label_key   = $element->args['label_key'] ?? __( 'Key', 'd4plib' );
+			$label_value = $element->args['label_value'] ?? __( 'Value', 'd4plib' );
+
+			echo '<div class="d4p-element-pair-single">';
+			echo '<div><span>' . $label_key . '</span></div>';
+			echo '<div><span>' . $label_value . '</span></div>';
+			echo '<span>&nbsp;</span>';
+			echo '</div>';
+		}
 
 		$i = 1;
 		foreach ( $value as $key => $val ) {
@@ -1044,13 +1050,32 @@ class Render {
 					'key'   => $key,
 					'value' => $val,
 				),
-				$element
+				$element,
+				false,
+				$layout
 			);
 			$i ++;
 		}
 
-		echo '<a role="button" class="button-primary" href="#">' . esc_html( $element->args['label_button_add'] ) . '</a>';
+		$this->_pair_element(
+			$name_base . '[' . $i . ']',
+			$id_base . '_' . $i,
+			$i,
+			array(
+				'key'   => '',
+				'value' => '',
+			),
+			$element,
+			false,
+			$layout
+		);
+
+		echo '</div>';
+
+		echo '<a role="button" class="button-primary" href="#">' . KSES::strong( $element->args['label_button_add'] ?? __( 'Add New', 'd4plib' ) ) . '</a>';
 		echo '<input type="hidden" value="' . esc_attr( $i ) . '" class="d4p-next-id" />';
+
+		echo '</div>';
 	}
 
 	protected function draw_expandable_raw( Element $element, $value, $name_base, $id_base = '' ) {