+++ dev4press/core/quick/BBP.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Name:    Dev4Press\v46\Core\Quick\BBP
- * Version: v4.6
+ * Name:    Dev4Press\v47\Core\Quick\BBP
+ * Version: v4.7
  * Author:  Milan Petrovic
  * Email:   support@dev4press.com
  * Website: https://www.dev4press.com/
@@ -25,7 +25,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-namespace Dev4Press\v46\Core\Quick;
+namespace Dev4Press\v47\Core\Quick;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -132,6 +132,14 @@ class BBP {
 		return $forums;
 	}
 
+	public static function is_user_moderator( int $user_id ) : bool {
+		return WPR::is_user_roles($user_id, bbp_get_moderator_role());
+	}
+
+	public static function is_user_keymaster( int $user_id ) : bool {
+		return WPR::is_user_roles($user_id, bbp_get_keymaster_role());
+	}
+
 	public static function is_current_user_moderator() : bool {
 		return WPR::is_current_user_roles( bbp_get_moderator_role() );
 	}
@@ -140,12 +148,13 @@ class BBP {
 		return WPR::is_current_user_roles( bbp_get_keymaster_role() );
 	}
 
-	/**
-	 * @depecated since 4.1, to be removed in 4.2. Use WPR::get_user_display_name() instead.
-	 */
-	public static function get_user_display_name( int $user_id = 0 ) : string {
-		_deprecated_function( __METHOD__, '4.1', 'WPR::get_user_display_name' );
+	public static function can_use_pretty_urls() : bool {
+		if ( function_exists( 'bbp_use_pretty_urls' ) ) {
+			return bbp_use_pretty_urls();
+		}
 
-		return WPR::get_user_display_name( $user_id );
+		global $wp_rewrite;
+
+		return $wp_rewrite->using_permalinks();
 	}
 }