+++ dev4press/core/quick/KSES.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * Name:    Dev4Press\v46\Core\Quick\Display
- * Version: v4.6
+ * Name:    Dev4Press\v47\Core\Quick\Display
+ * Version: v4.7
  * Author:  Milan Petrovic
  * Email:   support@dev4press.com
  * Website: https://www.dev4press.com/
@@ -25,7 +25,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>
  */
 
-namespace Dev4Press\v46\Core\Quick;
+namespace Dev4Press\v47\Core\Quick;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -154,7 +154,7 @@ class KSES {
 					'type'  => true,
 					'class' => true,
 				),
-				'span' => array(
+				'span'   => array(
 					'class' => true,
 				),
 			)
@@ -194,6 +194,7 @@ class KSES {
 					'class'        => true,
 					'style'        => true,
 					'title'        => true,
+					'alt'          => true,
 					'type'         => true,
 					'checked'      => true,
 					'readonly'     => true,
@@ -205,11 +206,13 @@ class KSES {
 					'max'          => true,
 					'step'         => true,
 					'size'         => true,
+					'minlength'    => true,
 					'maxlength'    => true,
 					'multiple'     => true,
 					'pattern'      => true,
 					'required'     => true,
 					'autocomplete' => true,
+					'autofocus'    => true,
 					'data-*'       => true,
 				),
 			)