+++ dev4press/core/ui/admin/PanelTools.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Dev4Press\v46\Core\UI\Admin;
+namespace Dev4Press\v47\Core\UI\Admin;
 
-use Dev4Press\v46\Core\Quick\KSES;
-use function Dev4Press\v46\Functions\panel;
+use Dev4Press\v47\Core\Quick\KSES;
+use function Dev4Press\v47\Functions\panel;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;