+++ forms/content-about-changelog.php
@@ -1,9 +1,9 @@
 <?php
 
-use Dev4Press\v46\Core\Helpers\Vendors;
-use Dev4Press\v46\Core\Quick\File;
-use Dev4Press\v46\Core\Quick\KSES;
-use function Dev4Press\v46\Functions\panel;
+use Dev4Press\v47\Core\Helpers\Vendors;
+use Dev4Press\v47\Core\Quick\File;
+use Dev4Press\v47\Core\Quick\KSES;
+use function Dev4Press\v47\Functions\panel;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;