+++ forms/content-settings-full.php
@@ -1,7 +1,7 @@
 <?php
 
-use Dev4Press\v46\Core\Options\Render;
-use function Dev4Press\v46\Functions\panel;
+use Dev4Press\v47\Core\Options\Render;
+use function Dev4Press\v47\Functions\panel;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -23,7 +23,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 		}
 
 		if ( isset( $obj['break'] ) ) {
-			echo panel()->r()->settings_break( $obj['break'], $obj['break-icon'] ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+			echo panel()->r()->settings_break( $obj['break'], $obj['break-icon'] ?? '', $obj['break-info'] ?? '' ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		}
 
 		echo panel()->r()->settings_group_break( $obj['title'], $obj['icon'] ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped